+++ src/components/Label/Label.stories.tsx
@@ -5,9 +5,9 @@ import colors from '../../enums/colors';
 import Label from './Label';
 import TextInput from '../TextInput';
 
-export const BasicLabel: Story = args => (
-  <Label {...args}>
-    <TextInput id={args.htmlFor} placeholder="placeholder" />
+export const BasicLabel: Story = ({ htmlFor, ...rest }) => (
+  <Label {...rest}>
+    <TextInput id={htmlFor} placeholder="placeholder" />
   </Label>
 );
 BasicLabel.args = {